reformatted code
split code into 2 classes so it's easier to work with
